Move sync directory into ownCloud
Rather than sync it through a git repository, we're going to sync
the settings via ownCloud. Because we're not syncing Mac apps
but primarily CLI apps with plain text files, there's less of a
possibility of conflicts when syncing. Bartender is one exception
b/c that has proven to be stable when syncing.